+++ app/Database/Migrations/2023-05-24-084121_Marketing.php
@@ -26,7 +26,7 @@ class Marketing extends Migration
                 'default' => null
             ),
             'actives' => array(
-                'type' => 'INT',
+                'type' => 'TINYINT',
                 'constraint' => 1,
                 'unsigned' => true,
                 'default' => 0
